backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.4.0
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==22.0.0
kombu==5.6.2
packaging==26.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
redis==5.0.8
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.7.0
